% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_updater}
\alias{bayes_updater}
\title{Bayesian Updating of Experimental Results}
\usage{
bayes_updater(est_1, se_1, est_2, se_2, graph = TRUE, ...)
}
\arguments{
\item{est_1}{Point estimate of Distribution 1}

\item{se_1}{Standard error on point estimate of Distribution 2}

\item{est_2}{Point estimate of Distribution 2}

\item{se_2}{Standard error on point estimate of Distribution 2}

\item{graph}{Boolean, should a graph be generated}

\item{...}{Any additional parameters that should be passed to ?graph_bayes}
}
\value{
vector of the posterior point estimate and SE. Optionally, graph
}
\description{
Generates posterior based on prior (Distribution 1) and likelihood (Distribution 2).
}
\examples{
bayes_updater(0, 0.1, 2.5, 0.1, x_min = -5, x_max = 5)
}
