% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate Outcome}
\usage{
estimate(
  outcome,
  covar,
  treat,
  family = "gaussian",
  type = "full",
  verbose = TRUE
)
}
\arguments{
\item{outcome}{Outcome data, as a numeric vector}

\item{covar}{Dataframe of covariates. Each column should be a separate covariate.}

\item{treat}{Treatment assignment}

\item{family}{Family for GLM}

\item{type}{`full` or `sparse`. `full` is a model with all covariates. `sparse` is a model with no covariates.}

\item{verbose}{Boolean. Should model output be printed?}
}
\value{
GLM model object
}
\description{
Estimate Outcome
}
\examples{
data(iris)
mm <- estimate(outcome = iris$Sepal.Length,
               covar = iris[, -1],
               treat = sample(0:1, nrow(iris), replace = TRUE))
}
