% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{graph_bayes}
\alias{graph_bayes}
\title{Graph Bayesian Updating}
\usage{
graph_bayes(est_1, se_1, est_2, se_2, est_pooled = NULL,
  se_pooled = NULL, x_min = -0.075, x_max = 0.075)
}
\arguments{
\item{est_1}{Point estimate of Distribution 1}

\item{se_1}{Standard error on point estimate of Distribution 2}

\item{est_2}{Point estimate of Distribution 2}

\item{se_2}{Standard error on point estimate of Distribution 2}

\item{est_pooled}{Point estimate of posterior, defaults to NULL (calculate automatically)}

\item{se_pooled}{Standard error of posterior, defaults to NULL (calculate automatically)}

\item{x_min}{Minimum value on x axis.}

\item{x_max}{Maximum value on x axis.}
}
\value{
ggplot2 object
}
\description{
Graphs posterior, prior (Distribution 1), and likelihood (Distribution 2). Requires ggplot2.
}
\examples{
graph_bayes(0, 0.01, 0.05, 0.01)
}
