% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup.R
\name{subgroup}
\alias{subgroup}
\title{Estimate Discrete Subgroup Effect}
\usage{
subgroup(outcome, subgroup, treat, family = "gaussian", verbose = TRUE)
}
\arguments{
\item{outcome}{Outcome data, as a numeric vector}

\item{subgroup}{Must be character vector}

\item{treat}{Treatment assignment}

\item{family}{Family for GLM}

\item{verbose}{Boolean. Should model output be printed?}
}
\value{
GLM model object
}
\description{
Estimate Discrete Subgroup Effect
}
\examples{
data(iris)
mm <- estimate(outcome = iris$Sepal.Length,
               subgroup = iris$Species,
               treat = sample(0:1, nrow(iris), replace = TRUE))
}
