% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.R
\name{submarg}
\alias{submarg}
\title{Return subgroup margins}
\usage{
submarg(
  mod,
  subgrp_var,
  treat_var = "treatment",
  mod_df = mod$data,
  weight = mod$prior.weights,
  ...
)
}
\arguments{
\item{mod}{glm}

\item{subgrp_var}{Subgroup variable, character. The subgroup variable must be
a categorical variable}

\item{treat_var}{Treatment variable, character}

\item{mod_df}{Dataframe used to estimate model, defaults to mod$data}

\item{weight}{Weight parameter to pass to `modmarg::marg`, defaults to mod$prior.weights}

\item{...}{Other parameters passed to modmarg::marg}
}
\value{

}
\description{
Calculates subgroup margins by running on subset of data. Doesn't matter for
OLS but does matter, e.g., for logit.
}
\examples{
library(modmarg)
data(margex)
mod <- glm(outcome ~ treatment * yc + distance, data = margex, family = 'binomial')
submarg(mod, 'yc', 'treatment')
}
