% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg.R
\name{topmarg}
\alias{topmarg}
\title{Return Topline Margins}
\usage{
topmarg(mod, treat_var = "treatment", ...)
}
\arguments{
\item{mod}{glm model}

\item{treat_var}{Treatment variable, character}

\item{...}{Other parameters passed to modmarg::marg}
}
\value{

}
\description{
Basically returns marginal effects on the treatment variable
}
\examples{
library(modmarg)
data(margex)
mod <- glm(outcome ~ treatment, data = margex, family = 'binomial')
topmarg(mod, 'treatment')
}
