% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate Outcome}
\usage{
estimate(
  outcome,
  covar,
  treat,
  data,
  sub_var = NULL,
  family = gaussian,
  type = "top",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{outcome}{character, outcome column.}

\item{covar}{character, columns for covariates}

\item{treat}{character, treatment assignment}

\item{data}{data frame to use for analysis}

\item{sub_var}{character, subgroup variable}

\item{family}{Family for glm}

\item{type}{`top` or `sub` based on topline or subgroup.}

\item{verbose}{Boolean. Should model output be printed?}

\item{...}{other paramters passed to glm}
}
\value{
GLM model object
}
\description{
Estimate Outcome
}
\examples{
data(iris)
iris$treat <- sample(0:1, nrow(iris), replace = TRUE)
mm <- estimate(
  outcome = 'Sepal.Length',
  covar = 'Sepal.Width',
  treat = 'treat', data = iris)
}
