% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Randomization Functions}
\usage{
rand(covariates, treatments, probs = NULL, blk_size = NULL)
}
\arguments{
\item{covariates}{data.frame or matrix of covariates}

\item{treatments}{}
}
\value{

}
\description{
Randomization Functions
}
\examples{
cov_df <- data.frame(
  a = sample(1:5, 1000, replace = T),
  b = sample(letters[1:10], 1000, replace = T))
rand(cov_df,
     treatments = c('Control', 'Treat'),
     probs = c(0.25, 0.75),
     blk_size = 4)
rand(cov_df,
     treatments = c('Control', 'Treat'))
}
